requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
alembic
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
